requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-extensions>=3.2.0
django-q2>=1.4.0
python-decouple>=3.8
django-filter>=23.0
django-ratelimit>=4.1.0